feat(plugins): Phase 5 — extract the Scene system into the SharpMUSH.Plugins.Scene reference plugin
The whole @SCENE command / scene… function / migration / SCENE_ROOM flag /
game.scene.* bridge surface now ships from a standalone, dynamically-loaded DLL
(SharpMUSH.Plugins.Scene, built into plugins/scene/) via the Phase-1/2a plugin
seams, while the ISceneService graph storage stays in the three DB providers.

Moved into the plugin:
- @SCENE command (SceneCommandModule + Scene*Handlers) and scene… functions
  (SceneFunctions), converted from partial Commands/Functions into standalone
  classes that resolve IMediator/INotifyService/ISceneService/config from
  parser.ServiceProvider; surfaced as ICommandSource/IFunctionSource via
  PluginBase + the generator analyzer.
- SceneBroadcast (the game.scene.{id} publish), called by the @scene arms.
- Arango Migration_AddScenes (plugin ArangoMigrationAssembly) + the
  Memgraph/Surreal scene schema (Cypher/Surreal statement lists), as IMigrationSource.
- The SCENE_ROOM flag, as IFlagSource (PluginFlag).
- The game.scene.* NATS->SignalR leg (was NatsBridgeService.SubscribeSceneAsync),
  as IBridgeSubscriptionSource (forwards via the non-generic IHubContext +
  SendAsync, so the plugin needs no reference to GameHub/IGameHubClient).

ScenePlugin implements PluginBase + IMigrationSource + IFlagSource +
IBridgeSubscriptionSource, so it is correctly a non-unloadable load-once plugin.

Removed from the engine: SceneCommand/* + SceneFunctions.cs + SceneBroadcast;
Migration_AddScenes + the SCENE_ROOM seed (Arango) and the scene schema +
SCENE_ROOM flag (Memgraph/Surreal built-in batches); SubscribeSceneAsync.

Kept: ISceneService + Models/Scene/* + SceneEventMessage; the providers'
*.Scene.cs storage; the ISceneService DI cast in Startup; SceneController + the
client portal; the DatabaseConstants scene names.

Wiring: SharpMUSH.Plugins.Scene is in the solution but referenced by the host and
test host with ReferenceOutputAssembly=false; a CopyScenePlugin MSBuild target
drops its DLL+deps.json+plugin.json into plugins/scene/ (mirroring SamplePlugin).
Boot ordering already holds: PluginBootstrapService registers plugin
commands/functions before DefaultPackagesBootstrapService installs the bundled
scene softcode package and before @STARTUP; migration/flag/bridge contributions
apply earlier during container construction via PluginCatalog.Build.

The five existing Scene test classes pass unchanged with Scene as a plugin
(SceneServiceIntegrationTests, SceneCommandFunctionIntegrationTests,
SceneDbrefResolutionTests, SceneHttpControllerTests, ScenePackageTests), and the
plugin suite (PluginManagerTests, PluginUnloadTests, PluginLoaderIntegrationTests)
stays green. Docs updated (plugin-system.md §Phase 5, scene-system.md §Phase 7).
