+++ SharpMUSH.Plugins.Scene/Commands/SceneMemberHandlers.cs
@@ -2,7 +2,7 @@ using SharpMUSH.Library.DiscriminatedUnions;
 using SharpMUSH.Library.ParserInterfaces;
 using SharpMUSH.Library.Services.Interfaces;
 
-namespace SharpMUSH.Implementation.Commands.SceneCommand;
+namespace SharpMUSH.Plugins.Scene.Commands;
 
 /// <summary>
 /// Membership / focus switches: MEMBER, UNMEMBER, FOCUS, SHOWAS.
