+++ SharpMUSH.Plugins.Scene/plugin.json
@@ -1,0 +1,7 @@
+{
+  "id": "scene",
+  "version": "1.0.0",
+  "dependencies": [],
+  "priority": 0,
+  "minServerVersion": null
+}
