+++ SharpMUSH.sln
@@ -63,6 +63,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tools", "tools", "{07C2787E
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PackageValidator", "tools\PackageValidator\PackageValidator.csproj", "{6A47CCCE-9E7B-44A3-B562-78DD1A020F80}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpMUSH.Plugins.Scene", "SharpMUSH.Plugins.Scene\SharpMUSH.Plugins.Scene.csproj", "{A478F699-3EC8-4B56-B2E2-F66E0E437277}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -349,6 +351,18 @@ Global
 		{6A47CCCE-9E7B-44A3-B562-78DD1A020F80}.Release|x64.Build.0 = Release|Any CPU
 		{6A47CCCE-9E7B-44A3-B562-78DD1A020F80}.Release|x86.ActiveCfg = Release|Any CPU
 		{6A47CCCE-9E7B-44A3-B562-78DD1A020F80}.Release|x86.Build.0 = Release|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Debug|x64.Build.0 = Debug|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Debug|x86.Build.0 = Debug|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Release|x64.ActiveCfg = Release|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Release|x64.Build.0 = Release|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Release|x86.ActiveCfg = Release|Any CPU
+		{A478F699-3EC8-4B56-B2E2-F66E0E437277}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
